+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 This file will only list released and supported versions, usually skipping over very minor updates.
 
+Unreleased
+==========
+
+4.0.0
+=====
+
+* Adds support for --firefox auto, picks the default firefox profile automatically
+* Upgrades to Python 3.0. Python 2 is no longer supported
+* Switches from PyGtk to PyGObject.
+* AppIndicator is no longer supported, because it is Python 2 only
+
 3.0.0
 =====